Move Deleter interface out of impl package
