+++ subprojects/build-cache-packaging/src/main/java/org/gradle/caching/internal/packaging/impl/PackerDirectoryUtil.java
@@ -17,8 +17,8 @@
 package org.gradle.caching.internal.packaging.impl;
 
 import org.apache.commons.io.FileUtils;
+import org.gradle.internal.file.Deleter;
 import org.gradle.internal.file.TreeType;
-import org.gradle.internal.file.impl.Deleter;
 
 import java.io.File;
 import java.io.IOException;
