+++ subprojects/core/src/main/java/org/gradle/api/tasks/Sync.java
@@ -26,7 +26,7 @@ import org.gradle.api.internal.file.copy.FileCopyAction;
 import org.gradle.api.internal.file.copy.SyncCopyActionDecorator;
 import org.gradle.api.tasks.util.PatternFilterable;
 import org.gradle.api.tasks.util.PatternSet;
-import org.gradle.internal.file.impl.Deleter;
+import org.gradle.internal.file.Deleter;
 import org.gradle.internal.reflect.Instantiator;
 
 import javax.inject.Inject;
