+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/GlobalScopeServices.java
@@ -72,8 +72,8 @@ import org.gradle.internal.environment.GradleBuildEnvironment;
 import org.gradle.internal.event.ListenerManager;
 import org.gradle.internal.execution.history.changes.DefaultExecutionStateChangeDetector;
 import org.gradle.internal.execution.history.changes.ExecutionStateChangeDetector;
+import org.gradle.internal.file.Deleter;
 import org.gradle.internal.file.impl.DefaultDeleter;
-import org.gradle.internal.file.impl.Deleter;
 import org.gradle.internal.filewatch.DefaultFileWatcherFactory;
 import org.gradle.internal.filewatch.FileWatcherFactory;
 import org.gradle.internal.fingerprint.overlap.OverlappingOutputDetector;
