+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/ProjectScopeServices.java
@@ -78,8 +78,8 @@ import org.gradle.configuration.project.ProjectConfigurationActionContainer;
 import org.gradle.initialization.ProjectAccessListener;
 import org.gradle.internal.Factory;
 import org.gradle.internal.build.BuildStateRegistry;
+import org.gradle.internal.file.Deleter;
 import org.gradle.internal.file.PathToFileResolver;
-import org.gradle.internal.file.impl.Deleter;
 import org.gradle.internal.hash.FileHasher;
 import org.gradle.internal.hash.StreamHasher;
 import org.gradle.internal.instantiation.InstantiatorFactory;
