+++ subprojects/execution/src/test/groovy/org/gradle/internal/execution/steps/CacheStepTest.groovy
@@ -31,7 +31,7 @@ import org.gradle.internal.execution.UnitOfWork
 import org.gradle.internal.execution.caching.CachingDisabledReason
 import org.gradle.internal.execution.caching.CachingDisabledReasonCategory
 import org.gradle.internal.execution.caching.CachingState
-import org.gradle.internal.file.impl.Deleter
+import org.gradle.internal.file.Deleter
 
 class CacheStepTest extends StepSpec<IncrementalChangesContext> implements FingerprinterFixture {
     def buildCacheController = Mock(BuildCacheController)
