+++ subprojects/files/src/main/java/org/gradle/internal/file/impl/DefaultDeleter.java
@@ -16,6 +16,7 @@
 package org.gradle.internal.file.impl;
 
 import com.google.common.annotations.VisibleForTesting;
+import org.gradle.internal.file.Deleter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
