+++ subprojects/platform-native/src/main/java/org/gradle/nativeplatform/toolchain/internal/swift/IncrementalSwiftCompiler.java
@@ -19,7 +19,7 @@ package org.gradle.nativeplatform.toolchain.internal.swift;
 import org.gradle.api.internal.TaskOutputsInternal;
 import org.gradle.api.tasks.WorkResult;
 import org.gradle.api.tasks.WorkResults;
-import org.gradle.internal.file.impl.Deleter;
+import org.gradle.internal.file.Deleter;
 import org.gradle.language.base.internal.compile.Compiler;
 import org.gradle.language.base.internal.tasks.SimpleStaleClassCleaner;
 import org.gradle.nativeplatform.internal.CompilerOutputFileNamingSchemeFactory;
